Update README.md
The steps of our paradigm include gene collection from the database, filtering, and finding the linkage of genes for these diseases. The initial step assists in the analysis by decreasing the total number of genes. Preprocessing, filtering, and linkage of genes were obtained by using the R language. After these analyses, we found four common genes. The topological properties for the related genes were identified. On the basis of the topological properties, Co-Expression and Physical Interaction networks were obtained. For systematic understanding of molecular mechanisms, we designed three types of gene regulatory networks. Based on our analyses, a drug is designed for the targeted genes.